[FLINK-23213][python] Move class SimpleStateRequestHandler into a separate file
